Crate nrf51

Reexports

extern crate core as core;

extern crate kernel;

use core::prelude::v1::*;

pub use chip::NRF51;

Modules

aes
AES128-CTR Driver

ble_advertising_driver
BLE Capsule

chip

clock
The clock peripheral of the nRF51 series (chapter 13 of the nRF51 reference manual v3.0), which manages the low frequency and high frequency clocks. The low frequency clock drives the real time clock (RTC), while the high frequency clocks drive the timer system.

gpio
GPIO and GPIOTE (task and events) for the nRF51 series.

nvic

peripheral_interrupts

peripheral_registers

pinmux
An abstraction over the NRF51 pin multiplexer

radio
Radio/BLE Driver for nrf51dk

rtc

temperature
Temperature Sensor Driver for nrf51dk

timer
The nRF51822 timer system operates off of the high frequency clock (HFCLK) and provides three timers from the clock. Timer0 is tied to the radio through some hard-coded peripheral linkages (e.g., there are dedicated PPI connections between Timer0's compare events and radio tasks, its capture tasks and radio events).

trng
TRNG driver for nrf51dk

uart

Functions

init⚠