Module nrf51::aes

AES128-CTR Driver

Provides a simple driver for userspace applications to encrypt and decrypt messages using aes128-ctr mode on top of aes128-ecb

The initial counter configred according to the counter received from the user application. The capsule is invoked as follows: - the key has been configured - the entire buffer has been encrypted - the entire buffer has been decrypted

The buffer is also sliced in chips at the moment and some un-necessary static mut...

FIXME: - maybe move some stuff to capsule instead - INIT_CTR can be replaced with TakeCell - ECB_DATA must be a static mut [u8] and can't be located in the struct - PAYLOAD size is restricted to 128 bytes

Author: Niklas Adolfsson niklasadolfsson1@gmail.com Author: Fredrik Nilsson frednils@student.chalmers.se Date: April 21, 2017

Reexports

use chip;

use core::cell::Cell;

use kernel::common::take_cell::TakeCell;

use kernel::hil::symmetric_encryption::SymmetricEncryption;

use kernel::hil::symmetric_encryption::Client;

use nvic;

use peripheral_interrupts::NvicIdx;

use peripheral_registers::AESECB_REGS;

use peripheral_registers::AESECB_BASE;

Structs

AesECB

Statics

AESECB

ECB_DATA

INIT_CTR

Functions

ECB_Handler⚠