Module nrf51::pinmux

An abstraction over the NRF51 pin multiplexer

Controller drivers should use the Pinmux type (instead of a u32) for fields that determine which pins are used by the hardware. The board configuration should create Pinmuxs and pass them into controller drivers during initialization.

Reexports

use kernel::common::VolatileCell;

Structs

Pinmux
An opaque wrapper around a configurable pin.

Statics

USED_PINS