Module nrf51::trng

TRNG driver for nrf51dk

The TRNG generates 1 byte randomness at the time value in the interval 0 <= r <= 255

The capsule requires 4 bytes of randomness

The counter "done" ensures that 4 bytes of randomness have been generated before returning to the capsule.

A temporary array "randomness" is used to store the randomness until it is returned to the capsule

In the current implementation if done > 4 for some strange reason the random generation will be restarted

Author: Niklas Adolfsson niklasadolfsson1@gmail.com Author: Fredrik Nilsson frednils@student.chalmers.se Date: March 01, 2017

Reexports

use chip;

use core::cell::Cell;

use core::mem;

use kernel::hil::rng;

use kernel::hil::rng::Continue;

use nvic;

use peripheral_interrupts::NvicIdx;

use peripheral_registers::RNG_BASE;

use peripheral_registers::RNG_REGS;

Structs

Trng

TrngIter

Statics

TRNG

Functions

RNG_Handler⚠